+++ backend/authentication/views.py
@@ -39,9 +39,7 @@ class SignupView(APIView):
 
         if serializer.is_valid():
             reg_source = request.query_params.get("source", "web")
-            print("Let's see the souce: ", reg_source)
             user = serializer.save(registration_source=reg_source)
-            print("user register source: ", user.registration_source)
 
             # Generate OTP
             otp = OTPService.generate_otp()