+++ backend/ax_merchant_api/urls.py
@@ -16,7 +16,11 @@ Including another URLconf
 """
 
 from orders.views import OrderCompleteView
-from dispatcher.views import MerchantAPIKeyRequestOTPView, MerchantAPIKeyRetrieveView
+from dispatcher.views import (
+    MerchantAPIKeyRequestOTPView,
+    MerchantAPIKeyRetrieveView,
+    MerchantRequestAPIAccessView,
+)
 from django.contrib import admin
 from django.urls import path, include
 
@@ -40,6 +44,12 @@ urlpatterns = [
     path("api/whatsapp/", include("whatsapp_messaging.urls")),
     path("api/subscriptions/", include("subscriptions.urls")),
     
+    path(
+        "api/merchant/request-api-access/",
+        MerchantRequestAPIAccessView.as_view(),
+        name="merchant_request_api_access",
+    ),
+
     # Merchant API Key management
     path(
         "api/merchant/apikey/request-otp/",